Add a command to copy my last doctrine migration.
When I've created a migration, I always want to open it in my editor.
When I use `--editor-cmd=vim`, it opens a new window though, and doesn't
open it in my current vim instance.

To fix this, I've created this command. When the migration version is
copied to my clipboard, I can access it really fast using CtrlP.